[sh0d4n]: added cmake

+++ UniquePtr/sources/main.cc
@@ -4,7 +4,7 @@
 
 int main() {
     custom_classes::unique_ptr<std::string []> my_ptr_1 = new std::string [1000];
-    for (int i = 0; i < 1000; ++i) my_ptr_1[i] = "a" ;
+    for (size_t i = 0; i < 1000; ++i) my_ptr_1[i] = "a" ;
 
     custom_classes::unique_ptr<int> my_ptr_2 = new int {3};
     custom_classes::unique_ptr<int> my_ptr_3 = new int {7};
